Aggregate changes
Changing the aggregate to extend from an aggregate root which uses a
trait called replayable which signifies that the object can replay
events.
Changing the aggregate so that it does not need to be initialized and
all data is passed through the appropriate methods.
Moving some items into the aggregate root class.
Updating the user aggregate spec to pass based on these modifications.