Broadcast factory
Adding a broadcast factory including an error broadcaster for user
creation.
Updating the interfaces for events and commands to use Contract in their
names.
Updating the AggregateRoot to return a UuidInterface to be consistent.
Updating the command for creation of users to be CreateUserListener.
Adding in an example exception class and factory for user exceptions.